requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_select=2.1.0=gpu
absl-py=0.7.0=py36_0
astor=0.7.1=py36_0
attrs=19.1.0=py36_1
autograd=1.3=pypi_0
autograd-gamma=0.4.1=pypi_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bottleneck=1.3.1=py36hdd07704_0
c-ares=1.15.0=h7b6447c_1
ca-certificates=2019.10.16=0
certifi=2019.9.11=py36_0
cffi=1.12.3=py36h2e261b9_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.6.0=cuda9.0_0
cupti=9.0.176=0
cvxopt=1.2.0=py36h54607b7_1
cvxpy=1.0.15=py36_0
cvxpy-base=1.0.15=py36hf484d3e_0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
dill=0.3.0=py36_0
ecos=2.0.7.post1=py36h5f7f5f7_0
expat=2.2.6=he6710b0_0
fastcache=1.1.0=py36h7b6447c_0
fftw=3.3.8=h7b6447c_3
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.17.1=pypi_0
gast=0.2.2=py36_0
glib=2.56.2=hd408876_0
glpk=4.65=h3ceedfd_2
gmp=6.1.2=h6c8ec71_1
grpcio=1.16.1=py36hf8bcb03_1
gsl=2.4=h14c3975_4
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
ipython=7.7.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
jedi=0.13.3=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.1=py36_0
jupyter_core=4.5.0=py_0
keras-applications=1.0.6=py36_0
keras-base=2.2.4=py36_0
keras-gpu=2.2.4=0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=h4deb6c0_3
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.6.1=hd408876_0
libstdcxx-ng=8.2.0=hdf63c60_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
lifelines=0.22.4=pypi_0
markdown=3.0.1=py36_0
matplotlib=3.1.0=py36h5429711_0
metis=5.1.0=hf484d3e_4
mifs=0.0.1.dev0=py36_0
mkl=2019.1=144
mkl_fft=1.0.10=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
multiprocess=0.70.8=py36h14c3975_0
nbformat=4.4.0=py36_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numexpr=2.6.9=py36h9e4a6bb_0
numpy=1.16.2=py36h7e9f1db_0
numpy-base=1.16.2=py36hde5b4d6_0
openssl=1.1.1d=h7b6447c_3
osqp=0.5.0=py36h5f7f5f7_0
pandas=0.24.2=py36he6710b0_0
parso=0.5.0=py_0
patsy=0.5.1=py36_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pip=19.0.3=py36_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.6.1=py36he6710b0_0
ptyprocess=0.6.0=py36_0
pycparser=2.19=py36_0
pygments=2.4.2=py_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytorch=1.1.0=py3.6_cuda9.0.176_cudnn7.5.1_0
pytz=2019.1=py_0
pyyaml=3.13=py36h14c3975_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scikit-learn=0.20.2=pypi_0
scikit-survival=0.10=py36h1211aff_0
scipy=1.2.1=py36h7c811a0_0
scs=2.1.1.2=py36hb3ffb1f_1
seaborn=0.9.0=py36_0
setuptools=40.8.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
statsmodels=0.10.1=py36hdd07704_0
suitesparse=5.2.0=h9e4a6bb_0
tbb=2019.8=hfd86e86_0
tensorboard=1.12.2=py36he6710b0_0
tensorflow=1.12.0=gpu_py36he68c306_0
tensorflow-base=1.12.0=gpu_py36h8e0ae2d_0
tensorflow-gpu=1.12.0=h0d30ee6_0
termcolor=1.1.0=py36_1
tk=8.6.8=hbc83047_0
torchsummary=1.5.1=pypi_0
tornado=6.0.3=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
werkzeug=0.14.1=py36_0
wheel=0.33.1=py36_0
xlrd=1.2.0=py36_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
